change: separate user tests from system tests
