+++ .github/workflows/action.yaml
@@ -35,10 +35,10 @@ jobs:
           restore-keys: ${{ runner.os }}-go-
       - name: Test Windows
         if: matrix.os == 'windows-latest'
-        run: go test -v .\...
+        run: go test -v -tags system .\...
       - name: 'Test *nix'
         if: matrix.os != 'windows-latest'
         # The CI has an old version of Go by default.
         # setup-go installs the version we want, but
         # root won't have that binary in its `secure_path`.
-        run: sudo $(which go) test -v ./...
+        run: sudo $(which go) test -v -tags system ./...
